refactor(tools): separate table functions from utils
